choferesbaja bugfix

+++ app/Http/Controllers/ChoferesController.php
@@ -21,6 +21,11 @@ class ChoferesController extends Controller
 
 
     public static function  eliminarChofer(User $user){
+
+        if(isset($user->combi)){
+            return redirect()->to(route('profile', ['user'=> $user]))-> with('tienecombi',$user->name);
+
+        }
         echo $user;
         DB::table('users')->whereId($user->id)->delete();
        return redirect()->to(route('listarchoferes'))-> with('usuarioeliminado',$user->name);
